requirements.txt
anthropic==0.40.0
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
PyMuPDF>=1.22.0
Pillow==10.2.0
fitz
streamlit